requirements.txt
# requirements.txt - Updated with compatible versions

# PDF Processing
PyMuPDF==1.24.0
pdfplumber==0.10.3

# Vector Database & Embeddings
chromadb==0.4.22
openai==1.12.0
tiktoken==0.5.2
httpx==0.27.0

# LLM & Agent
anthropic==0.18.1
langchain==0.1.9
langchain-community==0.0.24

# Knowledge Graph
networkx==3.2.1
matplotlib==3.8.2

# Search & Retrieval
rank-bm25==0.2.2

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
tqdm==4.66.2

# Testing
pytest==8.0.0
pytest-cov==4.1.0

# API
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9

# UI (Streamlit)
streamlit==1.31.1

# Data Processing
pandas==2.2.0
numpy==1.26.4
pathlib==1.0.1   # optional (useful for compatibility)
typing-extensions==4.10.0